% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicatorsFrmTbl.R
\name{indicatorsFromTbl}
\alias{indicatorsFromTbl}
\alias{prepareIndicPops}
\title{indicatorsFromTbl}
\usage{
prepareIndicPops(dfIndicDefs, dfData, nHhs)
}
\arguments{
\item{dfIndicDefs}{dataframe containing the indicator definitions}

\item{dfData}{dataframe containing the data that will be indicatified}

\item{nHhs}{number of households in the dataset}
}
\description{
function that turns table objects to indicators objects
}
\details{
Date created: 2017-11-25
Owner: nova Institute (Reg# 1994/002614/08).
Functions to prepare type 1, type 2 and type 3 indicators from indicator definition tables.
Adds the 'popnm' var to dfIndicDefs and generates the lsPops object - needed by functions such as indicsFrmTbl_tp1
}
