% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicatorsToDataframe.R
\name{indicatorsToDataframe}
\alias{indicatorsToDataframe}
\title{indicatorsToDataframe}
\usage{
indicatorsToDataframe(indicators)
}
\arguments{
\item{indicators}{An object of class 'indicator'}
}
\description{
function that turns indicator objects to dataframe
}
