% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicatorsToExcel.R
\name{indicatorsToExcel}
\alias{indicatorsToExcel}
\title{indicatorsToExcel}
\usage{
indicatorsToExcel(indicators, outputdir)
}
\arguments{
\item{indicators}{An object of class 'indicator'}

\item{outputdir}{character vector of length 1, containing the path to which the excel must be saved}
}
\description{
function that turns indicator objects to excel document
}
