% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_indicatorDefTbls_DES.R
\name{indicDefs_tp2}
\alias{indicDefs_tp2}
\title{indicDefs_tp2}
\usage{
indicDefs_tp2(dfHh)
}
\arguments{
\item{dfHh}{The data frame/vector containing indicator definitions/descriptions for type 2 (quantitative) indicators}
}
\value{
data.frame
}
\description{
Generates the indicator definitions/descriptions for type 2 (quantitative) indicators
}
\details{
Generates the indicator definitions/descriptions for type 2 (quantitative) indicators
}
