% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_indicatorDefTbls_DES.R
\name{indicDefs_tp3}
\alias{indicDefs_tp3}
\title{indicDefs_tp3}
\usage{
indicDefs_tp3()
}
\description{
Generates the indicator definitions/descriptions for type 3 (indicator X indicator) indicators
}
\details{
Generates the indicator definitions/descriptions for type 3 (indicator X indicator) indicators
}
