% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicatorsFrmTbl.R
\name{indicsFrmTbl_tp1}
\alias{indicsFrmTbl_tp1}
\title{indicsFrmTbl_tp1}
\usage{
indicsFrmTbl_tp1(dfIndicDefs, dfData, srcpth, nHhs)
}
\arguments{
\item{dfIndicDefs}{dataframe containing the indicator definitions}

\item{dfData}{dataframe containing the data that will be indicatified}

\item{srcpth}{path to the script that will generate this indicator}

\item{nHhs}{number of households in the dataset}
}
\description{
function that creates indicators for type 1 (discreet) data values e.g 'yes'/'no'
}
\details{
Calculate type 1 indicators ('yes'/'no' vars)
}
