% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicatorsFrmTbl.R
\name{indicsFrmTbl_tp2}
\alias{indicsFrmTbl_tp2}
\title{indicsFrmTbl_tp2}
\usage{
indicsFrmTbl_tp2(dfIndicDefs, dfData, srcpth, bPlot = TRUE)
}
\arguments{
\item{dfIndicDefs}{dataframe containing the indicator definitions}

\item{dfData}{dataframe containing the data that will be indicatified}

\item{srcpth}{path to the script that will generate this indicator}

\item{nHhs}{number of households in the dataset}
}
\description{
function that creates indicators for type 2 (quantitative) data values
}
\details{
Calculate type 2 indicators (quantitative)
}
