% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicatorsFrmTbl.R
\name{indicsFrmTbl_tp3}
\alias{indicsFrmTbl_tp3}
\title{indicsFrmTbl_tp3}
\usage{
indicsFrmTbl_tp3(dfIndicDefs, indicators)
}
\arguments{
\item{dfIndicDefs}{dataframe containing the indicator definitions}

\item{indicators}{An environment or list of indicator objects to be used in the}
}
\description{
function that creates indicators for type 2 (indicator X indicator) data values
}
\details{
Calculate type 3 indicators (indicator X indicator)
}
