% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_indicator2.R
\name{make.indicator}
\alias{make.indicator}
\title{make.indicator}
\usage{
make.indicator(x, name = "", digits = 2, NAsToZero = FALSE, indicObj = NULL)
}
\arguments{
\item{x}{A numeric vector}

\item{name}{name of the indicator}

\item{digits}{number of decimals after the comma when rounding}

\item{NAsToZero}{should NA's be made Zero}

\item{indicObj}{indicator object}
}
\description{
creates an indicator object from a numeric vector
}
