% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_indicator2.R
\name{make.indicatorNn}
\alias{make.indicatorNn}
\title{make.indicatorNn}
\usage{
make.indicatorNn(
  df,
  name,
  indicVar,
  indicOpt,
  groupVar = NULL,
  indicObj = NULL
)
}
\arguments{
\item{df}{data.frame containing the data}

\item{name}{name of the indicator object}

\item{indicVar}{name of the variable that must be indicatified}

\item{indicOpt}{option that must be indicatified}

\item{groupVar}{group variable, which the data must be grouped by when summarised}

\item{indicObj}{the indicator object}
}
\description{
function to indicatorise non-numeric (Nn) variables
}
