% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_indicator2.R
\name{print.indicator}
\alias{print.indicator}
\title{print.indicator}
\usage{
\method{print}{indicator}(x, inclCreatedBy = FALSE, inclComments = TRUE)
}
\arguments{
\item{x}{An object of class 'indicator'}

\item{inclCreatedBy}{logical, if TRUE it includes the persons name who created it}

\item{inclComments}{logical, if TRUE it includes the comment from the creator}
}
\description{
function that prints indicator neatly
}
